Added in some code for delete functionality, but still not working.

+++ server/schemas/resolvers.js
@@ -1,22 +1,25 @@
 const { User, Category, Activity, ActivityType } = require("../models");
 const { signToken, AuthenticationError } = require("../utils/auth");
-const mongoose = require('mongoose');
-mongoose.set('strictPopulate', false);
+const mongoose = require("mongoose");
+mongoose.set("strictPopulate", false);
 const resolvers = {
   Query: {
     me: async (parent, args, context) => {
       console.log(context);
       if (context.user) {
-        const userData = await User.findOne({ _id: context.user._id }).populate('activities').populate({
-          path: 'activities',
-          populate: 'category'
-        }).populate({
-          path: "user.activities",
-          populate: {
-            path: 'activityType', 
-            model: 'ActivityType'
-          }
-        });
+        const userData = await User.findOne({ _id: context.user._id })
+          .populate("activities")
+          .populate({
+            path: "activities",
+            populate: "category",
+          })
+          .populate({
+            path: "user.activities",
+            populate: {
+              path: "activityType",
+              model: "ActivityType",
+            },
+          });
         return userData;
       }
       throw AuthenticationError;
@@ -40,24 +43,27 @@ const resolvers = {
     },
 
     activities: async () => {
-      return await Activity.find({}).populate('category').populate('activityType');
+      return await Activity.find({})
+        .populate("category")
+        .populate("activityType");
     },
 
     // Query to fetch a single activity by ID
     activity: async (parent, { _id }) => {
-      return await Activity.findById(_id).populate('category').populate('activityType');
+      return await Activity.findById(_id)
+        .populate("category")
+        .populate("activityType");
     },
 
     // Query to fetch all activity types
     activityTypes: async () => {
-      return await ActivityType.find({}).populate('user');
+      return await ActivityType.find({}).populate("user");
     },
 
     // Query to fetch a single activity type by ID
     activityType: async (parent, { _id }) => {
-      return await ActivityType.findById(_id).populate('user');
+      return await ActivityType.findById(_id).populate("user");
     },
-
   },
 
   Mutation: {
@@ -85,45 +91,80 @@ const resolvers = {
     },
 
     // Mutation to add a new activity
-    addActivity: async (parent, { when, duration, commentText, category, activityType }, context) => {
+    addActivity: async (
+      parent,
+      { when, duration, commentText, category, activityType },
+      context
+    ) => {
       console.log(when);
-      const activity = await Activity.create({ when, duration, commentText, category: category, activityType: activityType })
-      console.log(activity)
+      const activity = await Activity.create({
+        when,
+        duration,
+        commentText,
+        category: category,
+        activityType: activityType,
+      });
+      console.log(activity);
 
       if (context.user) {
-
         const userData = await User.findOneAndUpdate(
           { _id: context.user._id },
           { $push: { activities: activity._id } },
-          { new: true });
+          { new: true }
+        );
         return activity;
       }
       throw AuthenticationError;
     },
 
     // Mutation to update an existing activity
-    updateActivity: async (parent, { _id, when, duration, commentText, categoryId, activityTypeId }) => {
+    updateActivity: async (
+      parent,
+      { _id, when, duration, commentText, categoryId, activityTypeId }
+    ) => {
       const updatedActivity = await Activity.findByIdAndUpdate(
         _id,
-        { when, duration, commentText, category: categoryId, activityType: activityTypeId },
+        {
+          when,
+          duration,
+          commentText,
+          category: categoryId,
+          activityType: activityTypeId,
+        },
         { new: true }
-      ).populate('category').populate('activityType');
+      )
+        .populate("category")
+        .populate("activityType");
       return updatedActivity;
     },
 
     // Mutation to delete an activity
-    deleteActivity: async (parent, { _id }) => {
-      const deletedActivity = await Activity.findByIdAndDelete(_id);
-      if (!deletedActivity) {
-        throw new Error('Activity not found');
+    deleteActivity: async (parent, { _id }, context) => {
+      if (context.user) {
+        const deleteActivity = await Activity.findByIdAndDelete(_id);
+        if (!deleteActivity) {
+          throw new Error("Activity not found");
+        }
       }
       return deletedActivity;
     },
 
+    // Example code
+    // removeSkill: async (parent, { skill }, context) => {
+    //   if (context.user) {
+    //     return Profile.findOneAndUpdate(
+    //       { _id: context.user._id },
+    //       { $pull: { skills: skill } },
+    //       { new: true }
+    //     );
+    //   }
+    //   throw AuthenticationError;
+    // },
+
     // Mutation to add a new activity type
     addActivityType: async (parent, { actName, userId }) => {
       const activityType = await ActivityType.create({ actName, userId });
-      return activityType.populate('userId');
+      return activityType.populate("userId");
     },
 
     // Mutation to update an existing activity type
@@ -132,7 +173,7 @@ const resolvers = {
         _id,
         { actName, userId },
         { new: true }
-      ).populate('userId');
+      ).populate("userId");
       return updatedActivityType;
     },
 
@@ -140,11 +181,10 @@ const resolvers = {
     deleteActivityType: async (parent, { _id }) => {
       const deletedActivityType = await ActivityType.findByIdAndDelete(_id);
       if (!deletedActivityType) {
-        throw new Error('ActivityType not found');
+        throw new Error("ActivityType not found");
       }
       return deletedActivityType;
     },
-
   },
 };
